Android support
Validated with NDK version android-ndk-r9b. This also includes
support for ARMs SDIV and UDIV ISA extensions to ARMv7 so u32 and
s32 integer divisions are done in hardware.

To execute on platforms which do not have these instructions the
remove the line which specifies ARM_SUPPORT_IDIV from Android.mk.

Signed-off-by: Tony Lofthouse <lofthous@amazon.com>